Removed the old commands and renamed the `-new` to the old name
