+++ bot/commands/calc.js
@@ -1,5 +1,5 @@
 const fight = require('../util/fightEngine')
-const units = require('./units')
+const { getBothUnitsArray, getUnitFromArray } = require('../unit/use-cases')
 
 module.exports = {
     name: 'calc',
@@ -21,39 +21,38 @@ module.exports = {
             return replyData
         }
 
-        const unitsArray = units.getBothUnitsArray(argsStr).filter(x => x != '')
-
-        const defenderStr = unitsArray.pop()
-        const defenderArray = defenderStr.split(/ +/)
-        const attackers = []
-
-        const defender = units.getUnitFromArray(defenderArray, replyData)
-        defender.getOverride(defenderArray, replyData)
-
-        unitsArray.forEach(x => {
-            const attackerArray = x.split(/ +/).filter(y => y != '')
-            const attacker = units.getUnitFromArray(attackerArray, replyData)
-            attacker.getOverride(attackerArray, replyData)
-            if (attacker.att !== 0)
-                attackers.push(attacker)
-        })
-
-        if (attackers.length === 0)
-            throw 'You need to specify at least one unit with more than 0 attack.'
-
         try {
+            const unitsArray = getBothUnitsArray(argsStr).filter(x => x != '')
+
+            const defenderStr = unitsArray.pop()
+            const defenderArray = defenderStr.split(/ +/)
+            const attackers = []
+
+            const defender = getUnitFromArray(defenderArray, replyData)
+
+            unitsArray.forEach(x => {
+                const attackerArray = x.split(/ +/).filter(y => y != '')
+                const attacker = getUnitFromArray(attackerArray, replyData)
+                if (attacker.att !== 0)
+                    attackers.push(attacker)
+            })
+
+            if (attackers.length === 0)
+                throw 'You need to specify at least one unit with more than 0 attack.'
+
             replyData = await fight.calc(attackers, defender, replyData)
+
+
+            dbData.attacker = attackers.length
+            dbData.defender = defender.name
+            dbData.defender_description = defender.description
+
+            if (replyData.discord.fields.length > 0)
+                dbData.reply_fields = [replyData.discord.fields[0].value.toString(), replyData.discord.fields[1].value]
+
+            return replyData
         } catch (error) {
             throw error
         }
-
-        dbData.attacker = attackers.length
-        dbData.defender = defender.name
-        dbData.defender_description = defender.description
-
-        if (replyData.discord.fields.length > 0)
-            dbData.reply_fields = [replyData.discord.fields[0].value.toString(), replyData.discord.fields[1].value]
-
-        return replyData
     }
 };
